filters: prevent spurious cancellation callbacks from the gRPC error path (#1560)
Description: Platform filters could fail to correctly transition to a completed state due to the lack of a response body on gRPC error responses. This resulted in spurious cancellation callbacks when the filter was scheduled for destruction due to the invalid "incomplete" status of the stream.
Risk Level: Low
Testing: Integration, Local & CI

Signed-off-by: Mike Schore <mike.schore@gmail.com>
Signed-off-by: JP Simard <jp@jpsim.com>