filters: add resetIdleTimer filter callback (#1577)
Description: This may be useful if a filter stops iteration for an extended period of time, since stream timeouts will still apply. The callback may be called periodically to continue to indicate "activity" on the stream. Note the final call to Envoy is commented out in this PR as we need the upstream change to land as well. Uncommenting and integration tests will come in a subsequent update.
Risk Level: Moderate
Testing: Local

Signed-off-by: Mike Schore <mike.schore@gmail.com>
Signed-off-by: JP Simard <jp@jpsim.com>