http: switching from XFP to scheme (#17372)
Corrects all Envoy uses of ForwardedProto which actually want request URI over to :scheme

As a reminder, XFP indicates the encryption of the (original) downstream connection where :scheme is part of the URI and the resource requested. It's legal (though unusual) to request http:// urls over a TLS connection for HTTP/2. It's possible (if ill advised) to have an internal mesh forwarding https schemed requests in the clear.

Current uses of X-Forwarded-Proto are

in the HCM, clearing XFP from untrusted users (unchanged)
in the HCM, setting absent XFP based on downstream transport security (unchanged)
in the HCM setting absent :scheme to XFP (unchanged)
in buildOriginalUri, changing from using XFP to scheme (changed. new URIs should be based on original URIs not on transport security.
in the router, clearing default port based on XFP (unchanged)
in the router serving redirect URLs based on scheme (changed - used to be XFP but is now based on the scheme of the original URI)
in the router, applying SSL route redirect based on XFP (unchanged)
in the router, using :scheme for internal redirect url checks (changed - used to use XFP. new URIs should be based on original URI)
in the cache filter, using :scheme to serve content (changed we used to serve based on XFP but if http://foo.com/ differs from https://foo.com and the http version is requested over a TLS connection the http response should be served)
in oath2 serving redirect URLs based on scheme (changed this used to be based on SFP but URLs should be based on original URL scheme)
Risk Level: High
Testing: updated tests
Docs Changes: inline
Release Notes: inline
Runtime guard: envoy.reloadable_features.correct_scheme_and_xfp
Fixes #14587

Signed-off-by: Alyssa Wilk <alyssar@chromium.org>