emit events based on ENVOY_LOG_EVENT  (#1746)
Description: Wires up the EM custom logger such that we intercept logs with stable names emitted by
ENVOY_LOG_EVENT, translating them into an event passed to the event tracker.

To maintain the original behavior when a logger is not installed, we end up having to partially
reimplement the Stderr logger.

For integration testing, we add a new test filter that emits an event, verifying that the engine
properly emits this both with and without a logger added.

Risk Level: Medium
Testing: Integration test
Docs Changes: n/a
Release Notes: n/a

Signed-off-by: Snow Pettersen <snowp@lyft.com>
Signed-off-by: JP Simard <jp@jpsim.com>