thread: Set up MainThread and TestThread contexts as RAII for more robust isMainThread checking (#18200)
Resolves an issue where people were adding ASSERT(Thread::MainThread::isMainThread()) into their systems, when in at least some cases that test was not detecting anything (failed open). This change enforces that we always create the singleton. It also protects (somewhat) against inter-test leakage (I hate singletons) by using RAII to declare threads, and tracking nested instantiations of TestThread and MainThread, forcing them to be consistent.

I don't actually anticipate TestThread ever changing, or needing this nested behavior, since we can declare that in main(). However, the main thread can be re-spawned between tests, and I suspect that some tests (fuzzing?) don't block till the previous test is completely quiesced. Previous to this PR that would lead to undefined behavior, and in this PR due to the nested tracking, it should ASSERT so that if such a leakage does occur, we'll learn about it.

Signed-off-by: Joshua Marantz <jmarantz@google.com>