[Apple] Guess string encoding when creating an NSString with UTF8 fails (#1891)
Description: We're seeing issues in the wild where some headers are being dropped because the `envoy_data` values convert to a nil `NSString` value. It's possible these headers aren't encoded in valid UTF-8. So in the case where the string conversion fails, we can attempt to have `NSString` guess an encoding. I've added a log with some debugging breadcrumbs in the case where a string could be created with a fallback encoding, and in the case where that fallback fails as well.
Risk Level: Low. I suspect that most string conversions succeed, so the fallback path should be rarely hit.
Testing: Existing header tests and ran the Hello World sample apps and confirmed that header conversion work as expected.
Fixes: https://github.com/envoyproxy/envoy-mobile/issues/1825 (possibly)

Signed-off-by: JP Simard <jp@jpsim.com>