[examples] Add Swift async/await example app (#1908)
I built this while debugging an issue a user had reported, and thought it would be useful to keep as an example app.

This example showcases a few integrations that weren't being exercised elsewhere I could find:

* Swift 5.5's new async/await & structured concurrency features
* Concurrent uploads to https://httpbin.org
* Starting streams on concurrent GCD queues
* SwiftUI interface

Finally, also add a CI job to validate that this example builds and logs the expected output.

Signed-off-by: JP Simard <jp@jpsim.com>