filter: Add scheme field to cache keys and deprecate clear_http (#19241)
Commit Message: Add scheme field to Envoy.Extensions.HttpFilters.Cache.Key and deprecate clear_http. The current state is confusing because the clear_http field is not marked "optional", so it doesn't have a "has" presence test, which means that if protocol information is ignored and the field isn't set, users of the proto will interpret clear_http=false as SSL=true.
Additional Description: An alternative would be to just add "optional" to the field, but the field is also confusing because "clear_http" sounds like an action where we "clear" an HTTP field. Also, adding presence tracking to an existing field may not be safe: If the field is set to its default value and passes through an old binary, the field will be dropped and any newer binaries downstream will think the field is unset.
Risk Level: Low
Testing: N/A
Docs Changes: N/A
Release Notes: N/A (the CacheFilter is still in Alpha)
Platform Specific Features: N/A

Signed-off-by: Edin Kadric <edin@google.com>