grpc_http1_bridge: support for prepending and trimming the grpc frame header (#19331)
This is an implementation based on the discussion in #19198.

When application/x-protobuf is used the filter will prepend the payload with the gRPC header, as defined by the wire format. Before returning the response the frame is also trimmed. This functionality is behind a configuration setting.

Risk Level: Medium
Testing: Unit tests
Docs Changes: Added
Fixes #19198

Signed-off-by: Stefan Puhlmann <stefanp@spotify.com>