GoogleGrpc: Always call Finish(), even if the stream is cancelled. (#19723)
The gRPC API has always required this, but until now, nothing bad has happened if the application failed to do so. However, an upcoming change in gRPC (grpc/grpc#28548) will enable transparent retries by default, at which point failing to call Finish() will cause a memory leak and/or connection leak. This PR fixes Envoy to use the gRPC API properly, so that it will not see this problem when it starts using a version of gRPC that supports transparent retries.

Risk Level: Low
Testing: Performed internal testing at Google with the gRPC change

Signed-off-by: Mark D. Roth <roth@google.com>