network: allow transport socket to intercept `connect` (#20213)
Currently a transport socket does not have the ability to intercept a `connect` call made in the
context of a request. In use cases which needs cluster context to make a connection, (for example
authenticating a connection's address tuple along with the backend service's FQN) we would need the
cluster context (for example service FQN) along with the connection. This can be done if we allow
the cluster's transport socket to intercept `connect`.

Signed-off-by: Jojy George Varghese <jojy_varghese@apple.com>
Co-authored-by: Greg Greenway <ggreenway@apple.com>