[iOS/macOS] Fix reachability (#2114)
Enable path monitor reachability state observation on both iOS and macOS.

It was disabled in https://github.com/envoyproxy/envoy-mobile/pull/2052 because
we didn't have a minimum macOS version set but this API is available as of
10.14.

It seems the `@platforms//os:macos` condition meant that the host was macOS, not
that the compilation target was macOS.

Risk Level: Low, re-enabling something that was enabled until fairly recently.
Testing: Validated that preferred network changes are logged when toggling WiFi
on/off in the Swift example apps on a physical iPhone. Also confirmed that the
library still builds for macOS with
`./bazelw build //library/objective-c:envoy_engine_objc_lib --config=macos`.
Release Notes: Added.

Signed-off-by: JP Simard <jp@jpsim.com>