bazel: set inmemory remote exec flags globally
I asked around and it sounds like these flags have no downside, but when
we only set them for remote exec, if you swapped between using remote
exec and then not using it, it would invalidate your build and you'd
rebuild the world. I'm hoping we can fix this rough edge in bazel but in
the meantime it should be safe to set them globally.

https://github.com/bazelbuild/bazel/pull/15203
https://github.com/bazelbuild/bazel/pull/15204
Signed-off-by: Keith Smiley <keithbsmiley@gmail.com>
Signed-off-by: JP Simard <jp@jpsim.com>