admin: factor out stats parameter parsing into a new class. (#20950)
The admin stats functions have a growing number of parameters (format, filter, histogram mode, usedonly), and we will want to add more. This PR is a straight refactor that packages them all up in a StatsParams object to make it easier to pass them around.

In particular, we'd like to add an alternative to the filter= query-param that uses RE2 rather than std::regex as the latter is extremely slow when run over 1M stats. That functional enhancement will be in a separate PR with doc changes & benchmark results.

Signed-off-by: Joshua Marantz <jmarantz@google.com>