skywalking: set peer with tag value of peer.address (#21137)
Peer is a important attribution of skywalking span. But in the previous skywalking tracer's implementation, it never be set correctly. This PR will set peer of skywalking span with that tag value of `peer.address`.

This PR consist of two part:
1. Set `peer.address` tag when we calling the `finalizeUpstreamSpan`.
2. Set peer of skywalking span base on the value of `peer.address` tag.

Risk Level: Low.
Testing: Added
Docs Changes: n/a.
Release Notes: n/a.
Platform Specific Features: n/a.

Signed-off-by: wbpcode <wangbaiping@corp.netease.com>
Co-authored-by: José Carlos Chávez <jcchavezs@gmail.com>