Add unused BalsaParser. (#21719)
Add unused BalsaParser.

Add first draft of BalsaParser implementation, which is currently
unused.  This fails a large number of tests and will require a lot of
work until it can be wired up to ConnectionImpl.

Also make Parser::getStatus() const.  That change is too trivial for a
separate PR.

Tracking issue: #21245

Signed-off-by: Bence Béky <bnc@google.com>