connectivity_manager: rename/refactor from Network::Configurator (#2401)
Description: This is intended to better capture the role and responsibilities of this class in Envoy Mobile. It still "configures connections", but more generally, it holds library/device connectivity state and provides mechanisms to leverage it.
Risk Level: Low
Testing: Updated tests.
Docs Changes: N/A
Release Notes: N/A

Signed-off-by: Mike Schore <mike.schore@gmail.com>
Signed-off-by: JP Simard <jp@jpsim.com>