docs: update find_class documentation (#2495)
Description: Update the documentation of `find_class` method as it seems to accept `/` and `.` and using `/` should be less confusing as this is what `env->findClass` expects.
Risk Level: None
Testing: Manual confirmed that the current implementing of tagging works and it does uses `/` as opposed to `.`.
Docs Changes: N/A
Release Notes: N/A

Signed-off-by: Rafał Augustyniak <Augustyniak@users.noreply.github.com>
Signed-off-by: JP Simard <jp@jpsim.com>