key_value: Use known registered name to create global PlatformInterfaceImpl (#2551)
The PlatformKeyValueStoreFactory used to take a PlatformInterfaceImpl where the registered KV store was found in the API registry and used to bridge the PlatformKeyValueStore with the native underlying store (e.g. SharedPrefs on Android).

However, registered factory classes only have their default constructor invoked, so the current PlatformKeyValueStoreFactory constructor taking in the PlatformInterfaceImpl would not have worked.

Instead, we use a globally known registered API name to fetch the registered KV store, and use it to create a static global PlatformInstanceImpl.

Signed-off-by: Ali Beyad <abeyad@google.com>
Signed-off-by: JP Simard <jp@jpsim.com>