build: Upgrade Abseil library to July 2022 release (#2580)
When enabling google_grpc in Envoy Mobile (which is off by default), the
build fails because Google gRPC depends on
absl/functional:any_invocable, which is not part of the Abseil library
that Envoy Mobile is on.

It turns out that Envoy Mobile's Abseil library is from a June 2021
release, so this commit upgrades the Abseil library used by Envoy Mobile
to the same one used by Envoy, which is from the July 5, 2022 release.

Signed-off-by: Ali Beyad <abeyad@google.com>
Signed-off-by: JP Simard <jp@jpsim.com>