Fix msan failure in the GrpcJsonTranscoderFilterTest.ResponseBodyExceedsBufferLimit (#23457)
msan triggers use of uninitialized value in the GrpcJsonTranscoderFilterTest.ResponseBodyExceedsBufferLimit test when a new slice with size == 0 is added to the buffer. Trigger point is just passing uninitialized value to the add() call. With size == 0 the add() call is a no-op.

Risk Level: Low
Testing: Unit Tests
Docs Changes: N/A
Release Notes: N/A
Platform Specific Features: N/A

Signed-off-by: Yan Avlasov <yavlasov@google.com>