Stricter boundary checks for Duration when it is converted to std::chrono::milliseconds (#23529)
Internally protobuf first converts seconds to nanoseconds before converting to milliseconds. This can overflow if the default bound checks are applied to the seconds value.

This can potentially be a breaking change for configuration that erroneously supplied values higher than 9223372036, since Envoy will now be rejecting such config. Before it was accepting it and using the "wrapped around" value.

This case is also verified by the test/server/server_corpus/clusterfuzz-testcase-server_fuzz_test-5988544525893632 corpus, however the asan sanitizer needs to be of a newer version to catch the error.

Signed-off-by: Yan Avlasov <yavlasov@google.com>