tls: remove stale comment (#23846)
Both things the comment says are inaccurate: After #23320, this
mechanism no longer uses the X509_STORE callback, but a flag.
After #21417, Envoy is also not (always) using
SSL_CTX_set_cert_verify_callback, but a different one.

Just remove the comment altogether, as it doesn't seem to provide any
value. All it's saying is that, because Envoy still uses `store` at the
end of the day, configuring things on `store` does stuff. But the whole
function configures things on `store`, so it's not specific to that
block anyway.

Signed-off-by: David Benjamin <davidben@google.com>