Support allowlisting request headers for both gRPC and HTTP authorization servers (#22877)
Add support for allowlisting request headers included in the check request to the authorization server, be it HTTP or gRPC (currently, this is supported for HTTP only). This patch deprecates the current allowed_headers config field (on the authorization_request message) and introduces a new one on the ext_authz message. To note is that the default behaviour (i.e. no config specified) differs based on the type of authorization server (this is to maintain backwards compatibility).

Signed-off-by: rulex123 <erica.manno@gmail.com>