android proxy: fix initial proxy settings state (#2709)
Description: Upon investigation of multiple sessions with active proxy settings it seems to be highly likely that our EM proxy logic did not engage in all of the aforementioned sessions. That appears to be a direct result of the way our logic for registering to `PROXY_CHANGE` intent works. The implementation assumes that upon subscription to the intent we always receive an initial intent with the state of proxy settings. That does not seem to be guaranteed by Android as `PROXY_CHANGE` intent is not documented to be sticky (a term used by Android) and whether it's sticky or not may depend on manufacturer / type of the proxy update.
Risk Level: Low, hidden behind a feature flag and disabled by default.
Testing: Integration tests
Docs Changes: N/A
Release Notes: N/A

Signed-off-by: Rafal Augustyniak <raugustyniak@lyft.com>
Signed-off-by: JP Simard <jp@jpsim.com>